hack: disable the ts2021 endpoint
We do this because the ts2021 endpoint in Headscale currently does not
support real WebSocket clients.

We tried to add support by enabling GET requests to the ts2021 endpoint,
but the connection is killed by some context cancelling before any data
is exchanged (except for the noise handshake).
Replacing a few contexts with context.Background() make it work after a
few tries, but never clean on the first try.
For now, we just fall back to the legacy protocol, that works.